+++ changelog.md
@@ -29,6 +29,7 @@ Here are the changes not related to the renamed API.
 * `phoenix::model_script::parse_binary` has been removed in favour of `phoenix::model_script::parse`
 * `phoenix::vobs::camera_lock_mode` has been removed in favour of `phoenix::vobs::sprite_alignment`
 * `phoenix::vob::camera_alignment` has been removed in favour of `phoenix::vob::sprite_camera_facing_mode`
+* `phoenix::symbol::{get,set}_{string,int,float}` now take a raw pointer as context.
 
 ### Features
 * [cfec0051] XZEN-encoded worlds are now supported. Thanks, @ThielHater!